feat: add verify version of getSession
Until now, there was one version of getSession, which was verifying the driver connection.  That has
3 drawbacks: (1) getSession had to be async; (2) the verification was an extra call, that may affect
performance speed in some applications; (3) the need for the DB certifications is an issue for CI.
To make a unit test where getSession is called but the session is not used, you'd have to stub
getSession.  The current approach is to allow an async `getSessionVerify`, which is the old version,
and to make `getSession` sync.

BREAKING CHANGE: When you run `getSession()` now, it will no longer verify the connection.  While it
is not a breaking change, you no longer need `await`, and `getSession` can now be called outside of
async functions, which can be very handy.